refactor: isolate responses terminal failure formatting
